continuing refactoring
