+++ server/src/graql/gremlin/RelationTypeInference.java
@@ -1,0 +1,166 @@
+package grakn.core.graql.gremlin;
+
+import com.google.common.collect.HashMultimap;
+import com.google.common.collect.ImmutableSet;
+import com.google.common.collect.Iterators;
+import com.google.common.collect.Multimap;
+import com.google.common.collect.Sets;
+import grakn.core.concept.Label;
+import grakn.core.concept.type.RelationType;
+import grakn.core.concept.type.Role;
+import grakn.core.concept.type.SchemaConcept;
+import grakn.core.concept.type.Type;
+import grakn.core.graql.gremlin.fragment.Fragment;
+import grakn.core.graql.gremlin.fragment.Fragments;
+import grakn.core.graql.gremlin.fragment.InIsaFragment;
+import grakn.core.graql.gremlin.fragment.InSubFragment;
+import grakn.core.graql.gremlin.fragment.LabelFragment;
+import grakn.core.graql.gremlin.fragment.OutRolePlayerFragment;
+import grakn.core.graql.gremlin.sets.EquivalentFragmentSets;
+import grakn.core.server.session.TransactionOLTP;
+import graql.lang.property.IsaProperty;
+import graql.lang.property.TypeProperty;
+import graql.lang.statement.Statement;
+import graql.lang.statement.Variable;
+
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Iterator;
+import java.util.Map;
+import java.util.Set;
+
+import static graql.lang.Graql.var;
+
+public class RelationTypeInference {
+    // infer type of relation type if we know the type of the role players
+    // add label fragment and isa fragment if we can infer any
+    public static Set<Fragment> inferRelationTypes(TransactionOLTP tx, Set<Fragment> allFragments) {
+
+        Set<Fragment> inferredFragments = new HashSet<>();
+
+        Map<Variable, Type> labelVarTypeMap = getLabelVarTypeMap(tx, allFragments);
+        if (labelVarTypeMap.isEmpty()) return inferredFragments;
+
+        Multimap<Variable, Type> instanceVarTypeMap = getInstanceVarTypeMap(allFragments, labelVarTypeMap);
+
+        Multimap<Variable, Variable> relationRolePlayerMap = getRelationRolePlayerMap(allFragments, instanceVarTypeMap);
+        if (relationRolePlayerMap.isEmpty()) return inferredFragments;
+
+        // for each type, get all possible relation type it could be in
+        Multimap<Type, RelationType> relationMap = HashMultimap.create();
+        labelVarTypeMap.values().stream().distinct().forEach(
+                type -> addAllPossibleRelations(relationMap, type));
+
+        // inferred labels should be kept separately, even if they are already in allFragments set
+        Map<Label, Statement> inferredLabels = new HashMap<>();
+        relationRolePlayerMap.asMap().forEach((relationVar, rolePlayerVars) -> {
+
+            Set<Type> possibleRelationTypes = rolePlayerVars.stream()
+                    .filter(instanceVarTypeMap::containsKey)
+                    .map(rolePlayer -> getAllPossibleRelationTypes(
+                            instanceVarTypeMap.get(rolePlayer), relationMap))
+                    .reduce(Sets::intersection).orElse(Collections.emptySet());
+
+            //TODO: if possibleRelationTypes here is empty, the query will not match any data
+            if (possibleRelationTypes.size() == 1) {
+
+                Type relationType = possibleRelationTypes.iterator().next();
+                Label label = relationType.label();
+
+                // add label fragment if this label has not been inferred
+                if (!inferredLabels.containsKey(label)) {
+                    Statement labelVar = var();
+                    inferredLabels.put(label, labelVar);
+                    Fragment labelFragment = Fragments.label(new TypeProperty(label.getValue()), labelVar.var(), ImmutableSet.of(label));
+                    inferredFragments.add(labelFragment);
+                }
+
+                // finally, add inferred isa fragments
+                Statement labelVar = inferredLabels.get(label);
+                IsaProperty isaProperty = new IsaProperty(labelVar);
+                EquivalentFragmentSet isaEquivalentFragmentSet = EquivalentFragmentSets.isa(isaProperty,
+                        relationVar, labelVar.var(), relationType.isImplicit());
+                inferredFragments.addAll(isaEquivalentFragmentSet.fragments());
+            }
+        });
+
+        return inferredFragments;
+    }
+
+    // find all vars with direct or indirect out isa edges
+    private static Multimap<Variable, Type> getInstanceVarTypeMap(
+            Set<Fragment> allFragments, Map<Variable, Type> labelVarTypeMap) {
+        Multimap<Variable, Type> instanceVarTypeMap = HashMultimap.create();
+        int oldSize;
+        do {
+            oldSize = instanceVarTypeMap.size();
+            allFragments.stream()
+                    .filter(fragment -> labelVarTypeMap.containsKey(fragment.start())) // restrict to types
+                    .filter(fragment -> fragment instanceof InIsaFragment || fragment instanceof InSubFragment) //
+                    .forEach(fragment -> instanceVarTypeMap.put(fragment.end(), labelVarTypeMap.get(fragment.start())));
+        } while (oldSize != instanceVarTypeMap.size());
+        return instanceVarTypeMap;
+    }
+
+    // find all vars representing types
+    private static Map<Variable, Type> getLabelVarTypeMap(TransactionOLTP tx, Set<Fragment> allFragments) {
+        Map<Variable, Type> labelVarTypeMap = new HashMap<>();
+        allFragments.stream()
+                .filter(LabelFragment.class::isInstance)
+                .forEach(fragment -> {
+                    // TODO: labels() should return ONE label instead of a set
+                    SchemaConcept schemaConcept = tx.getSchemaConcept(
+                            Iterators.getOnlyElement(((LabelFragment) fragment).labels().iterator()));
+                    if (schemaConcept != null && !schemaConcept.isRole() && !schemaConcept.isRule()) {
+                        labelVarTypeMap.put(fragment.start(), schemaConcept.asType());
+                    }
+                });
+        return labelVarTypeMap;
+    }
+
+    private static Multimap<Variable, Variable> getRelationRolePlayerMap(
+            Set<Fragment> allFragments, Multimap<Variable, Type> instanceVarTypeMap) {
+        // relation vars and its role player vars
+        Multimap<Variable, Variable> relationRolePlayerMap = HashMultimap.create();
+        allFragments.stream().filter(OutRolePlayerFragment.class::isInstance)
+                .forEach(fragment -> relationRolePlayerMap.put(fragment.start(), fragment.end()));
+
+        // find all the relation requiring type inference
+        Iterator<Variable> iterator = relationRolePlayerMap.keySet().iterator();
+        while (iterator.hasNext()) {
+            Variable relation = iterator.next();
+
+            // the relation should have at least 2 known role players so we can infer something useful
+            if (instanceVarTypeMap.containsKey(relation) ||
+                    relationRolePlayerMap.get(relation).size() < 2) {
+                iterator.remove();
+            } else {
+                int numRolePlayersHaveType = 0;
+                for (Variable rolePlayer : relationRolePlayerMap.get(relation)) {
+                    if (instanceVarTypeMap.containsKey(rolePlayer)) {
+                        numRolePlayersHaveType++;
+                    }
+                }
+                if (numRolePlayersHaveType < 2) {
+                    iterator.remove();
+                }
+            }
+        }
+        return relationRolePlayerMap;
+    }
+
+    private static void addAllPossibleRelations(Multimap<Type, RelationType> relationMap, Type metaType) {
+        metaType.subs().forEach(type -> type.playing().flatMap(Role::relations)
+                .forEach(relationType -> relationMap.put(type, relationType)));
+    }
+
+    private static Set<Type> getAllPossibleRelationTypes(
+            Collection<Type> instanceVarTypes, Multimap<Type, RelationType> relationMap) {
+
+        return instanceVarTypes.stream()
+                .map(rolePlayerType -> (Set<Type>) new HashSet<Type>(relationMap.get(rolePlayerType)))
+                .reduce(Sets::intersection).orElse(Collections.emptySet());
+    }
+}
